+++ src/graphql/resolvers/user.resolver.ts
@@ -1,6 +1,7 @@
 import { Profile } from '@/types/profiles.types';
 import { userService } from '../../service/user.service';
 import { argsType, AuthenticatedUser, User } from '../../types/users.types';
+import { Context } from '../context';
 
 const resolvers = {
   User: {
@@ -9,17 +10,24 @@ const resolvers = {
     },
   },
   Query: {
-    getAllUsers: async (): Promise<User[] | undefined> => {
+    getAllUsers: async (
+      _obj: unknown,
+      _args: argsType,
+      context: Context,
+    ): Promise<User[] | undefined> => {
+      context.validateAdmin();
       return await userService.getAllUsers();
     },
     getUserByParams: async (
-      _parent: unknown,
+      _obj: unknown,
       args: argsType,
+      context: Context,
     ): Promise<User | undefined> => {
+      context.validateUserAccess(args!.filter!.id!);
       return await userService.getUserByParams(args?.filter);
     },
     login: async (
-      _parent: unknown,
+      _obj: unknown,
       args: argsType,
     ): Promise<AuthenticatedUser | undefined> => {
       return await userService.login(args?.input);
@@ -27,25 +35,31 @@ const resolvers = {
   },
   Mutation: {
     createUser: async (
-      _parent: unknown,
+      _obj: unknown,
       args: argsType,
+      context: Context,
     ): Promise<User | undefined> => {
+      context.validateAdmin();
       return await userService.createUser(args?.input);
     },
     updateUser: async (
-      _parent: unknown,
+      _obj: unknown,
       args: argsType,
+      context: Context,
     ): Promise<User | undefined> => {
+      context.validateUserAccess(args?.id);
       return await userService.updateUser(args?.id, args?.input);
     },
     deleteUser: async (
-      _parent: unknown,
+      _obj: unknown,
       args: argsType,
+      context: Context,
     ): Promise<string | undefined> => {
+      context.validateAdmin();
       return await userService.deleteUser(args?.id);
     },
     signUp: async (
-      _parent: unknown,
+      _obj: unknown,
       args: argsType,
     ): Promise<User | undefined> => {
       return await userService.signUp(args?.input);
